common: move hex_decode() to parsing.c and remove parse_hex()
parse_hex() and hex_decode() both implement a hex decoder, we had an
additional hex_decode() in yhauth as well that also gets removed and use
the one in common.
