+++ common/parsing.h
@@ -19,6 +19,7 @@
 
 #include <stdbool.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include "../common/platform-config.h"
 
 // NOTE(adma): those utility functions do not link against libyubihsm
@@ -35,5 +36,6 @@ enum input_visibility { VISIBLE, HIDDEN_CHECKED, HIDDEN_UNCHECKED };
 bool YH_INTERNAL read_string(const char *name, char *str_buf,
                              size_t str_buf_len,
                              enum input_visibility visibility);
+bool YH_INTERNAL hex_decode(const char *in, uint8_t *out, size_t *len);
 
 #endif
