+++ common/util.c
@@ -15,9 +15,7 @@
  */
 
 #include <ctype.h>
-#include <errno.h>
 #include <string.h>
-#include <limits.h>
 
 #include <openssl/bio.h>
 #include <openssl/evp.h>
@@ -557,41 +555,6 @@ bool base64_decode(const char *in, uint8_t *out, size_t *len) {
   }
 }
 
-bool hex_decode(const char *in, uint8_t *out, size_t *len) {
-  int pos = 0;
-  size_t in_len = strlen(in);
-  if (in_len > 0 && in[in_len - 1] == '\n') {
-    in_len--;
-  }
-  if (in_len > 0 && in[in_len - 1] == '\r') {
-    in_len--;
-  }
-  if (in_len % 2 != 0) {
-    return false;
-  } else if (in_len / 2 > *len) {
-    return false;
-  }
-
-  for (size_t i = 0; i < in_len / 2; i++) {
-    char *endptr = NULL;
-    char buf[3] = {0};
-    long num;
-    errno = 0;
-
-    buf[0] = in[pos];
-    buf[1] = in[pos + 1];
-    num = strtol((const char *) buf, &endptr, 16);
-    if ((errno == ERANGE && (num < 0 || num > UCHAR_MAX)) ||
-        (errno != 0 && num == 0) || *endptr != '\0') {
-      return false;
-    }
-    out[i] = (uint8_t) num;
-    pos += 2;
-  }
-  *len = in_len / 2;
-  return true;
-}
-
 bool write_file(const uint8_t *buf, size_t buf_len, FILE *fp, format_t format) {
 
   const uint8_t *p = buf;
@@ -744,30 +707,4 @@ bool split_hmac_key(yh_algorithm algorithm, uint8_t *in, size_t in_len,
   *out_len = 2 * block_size;
 
   return true;
-}
-
-size_t parse_hex(const char *hex, size_t hex_len, uint8_t *parsed) {
-
-  size_t j = 0;
-
-  for (size_t i = 0; i < hex_len; i += 2) {
-    if (isxdigit(hex[i]) == 0 || isxdigit(hex[i + 1]) == 0) {
-      break;
-    }
-
-    if (isdigit(hex[i])) {
-      parsed[j] = (hex[i] - '0') << 4;
-    } else {
-      parsed[j] = (tolower(hex[i]) - 'a' + 10) << 4;
-    }
-
-    if (isdigit(hex[i + 1])) {
-      parsed[j] |= (hex[i + 1] - '0');
-    } else {
-      parsed[j] |= (tolower(hex[i + 1]) - 'a' + 10);
-    }
-
-    j++;
-  }
-  return j;
 }
